# web/src/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorDetail": {
        "additionalProperties": false,
        "properties": {
          "location": {
            "description": "Where the error occurred, e.g. 'body.items[3].tags' or 'path.thing-id'",
            "type": "string"
          },
          "message": {
            "description": "Error message text",
            "type": "string"
          },
          "value": {
            "description": "The value at the given location"
          }
        },
        "type": "object"
      },
      "ErrorModel": {
        "additionalProperties": false,
        "properties": {
          "detail": {
            "description": "A human-readable explanation specific to this occurrence of the problem.",
            "examples": [
              "Property foo is required but is missing."
            ],
            "type": "string"
          },
          "errors": {
            "description": "Optional list of individual error details",
            "items": {
              "$ref": "#/components/schemas/ErrorDetail"
            },
            "type": "array"
          },
          "instance": {
            "description": "A URI reference that identifies the specific occurrence of the problem.",
            "examples": [
              "https://example.com/error-log/abc123"
            ],
            "format": "uri",
            "type": "string"
          },
          "status": {
            "description": "HTTP status code",
            "examples": [
              400
            ],
            "format": "int64",
            "type": "integer"
          },
          "title": {
            "description": "A short, human-readable summary of the problem type. This value should not change between occurrences of the error.",
            "examples": [
              "Bad Request"
            ],
            "type": "string"
          },
          "type": {
            "default": "about:blank",
            "description": "A URI reference to human-readable documentation for the error.",
            "examples": [
              "https://example.com/errors/example"
            ],
            "format": "uri",
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "UpdateStatus": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "type": "string"
          },
          "requested": {
            "type": "boolean"
          },
          "state": {
            "type": "string"
          },
          "supported": {
            "type": "boolean"
          },
          "version": {
            "type": "string"
          }
        },
        "required": [
          "state",
          "version",
          "requested",
          "supported"
        ],
        "type": "object"
      },
...
      "WarpPatchInputBody": {
        "additionalProperties": false,
        "properties": {
          "enabled": {
            "type": "boolean"
          },
          "license": {
            "description": "Ключ WARP+ для зарегистрированного аккаунта",
            "maxLength": 40,
            "type": "string"
          },
          "routes": {
            "description": "Домены (example.com — вместе с поддоменами) и сети (104.16.0.0/13)",
            "items": {
              "type": "string"
            },
            "maxItems": 1000,
            "type": "array"
          }
        },
        "type": "object"
      },
...
      "WarpView": {
        "additionalProperties": false,
        "properties": {
          "configured": {
            "type": "boolean"
          },
          "enabled": {
            "type": "boolean"
          },
          "endpoint": {
            "type": "string"
          },
          "inbounds": {
            "description": "Подключения ноды, у которых весь трафик идёт через WARP",
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "ipv4": {
            "type": "string"
          },
          "ipv6": {
            "type": "string"
          },
          "plus": {
            "description": "Аккаунт WARP+",
            "type": "boolean"
          },
          "routes": {
            "description": "Домены и сети, которые идут через WARP у всех подключений ноды",
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "source": {
            "enum": [
              "register",
              "import",
              ""
            ],
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/WarpViewStatusStruct",
            "description": "Последняя проверка выхода через WARP с ноды; нет — нода недоступна"
          }
        },
        "required": [
          "configured",
          "enabled",
          "plus",
          "routes",
          "inbounds"
        ],
        "type": "object"
      },
      "WarpViewStatusStruct": {
        "additionalProperties": false,
        "properties": {
          "checked_at": {
            "format": "date-time",
            "type": "string"
          },
          "colo": {
            "type": "string"
          },
          "error": {
            "type": "string"
          },
          "ip": {
            "description": "Адрес, который видят сайты",
            "type": "string"
          },
          "ok": {
            "type": "boolean"
          },
          "warp": {
            "description": "on | plus | off — как отвечает Cloudflare",
            "type": "string"
          }
        },
        "required": [
          "ok",
          "checked_at"
        ],
        "type": "object"
      }
    },
    "securitySchemes": {
      "session": {
        "description": "Сессия админки + заголовок X-CSRF-Token на изменяющих запросах",
        "in": "cookie",
        "name": "__Host-mirai",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "REST API панели mirai. Все пути — под секретным адресом админки: https://\u003cпанель\u003e/\u003cсекретный путь\u003e/api/v1/…\n\nАдминка в браузере ходит с cookie сессии; изменяющие запросы тогда требуют заголовок `X-CSRF-Token` из `GET /auth/me`.\n\nОшибки — RFC 9457 (application/problem+json): `detail` — код ошибки, `errors[].message` — код по полю.",
    "title": "mirai",
    "version": "dev"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/nodes/{id}/update": {
      "get": {
        "operationId": "node-update-status",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "minimum": 1,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateStatus"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Статус обновления ноды",
        "tags": [
          "node"
        ]
      },
      "post": {
        "operationId": "request-node-update",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "minimum": 1,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Accepted"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Обновить ноду",
        "tags": [
          "node"
        ],
        "x-session-only": true
      }
    },
    "/api/v1/nodes/{id}/warp": {
      "delete": {
        "operationId": "delete-node-warp",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "minimum": 1,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Удалить WARP ноды",
        "tags": [
          "node"
        ]
      },
      "get": {
        "operationId": "get-node-warp",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "minimum": 1,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WarpView"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "WARP ноды",
        "tags": [
          "node"
        ]
      },
      "patch": {
        "operationId": "update-node-warp",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "minimum": 1,
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WarpPatchInputBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WarpView"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Включить WARP, списки доменов, ключ WARP+",
        "tags": [
          "node"
        ]
      }
    },
...
  },
  "security": [
    {
      "session": []
    }
  ]
}